requirements.txt
streamlit==1.28.1
openai==1.3.7
anthropic==0.7.8
google-generativeai==0.3.2
langchain==0.0.350
langchain-community==0.0.10
langchain-openai==0.0.2
faiss-cpu==1.7.4
pypdf2==3.0.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
